ml_service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
tensorflow>=2.12.0
numpy>=1.24.0
pillow>=10.0.0
python-multipart>=0.0.9
